.ipynb_checkpoints/requirements-checkpoint.txt
python-dotenv==0.15.0
rauth==0.7.2
pathlib==1.0.1
mysql-connector-python==8.0.23




